refactor: memberCoupon 패키지 이동
