+++ src/main/java/cart/repository/MemberCouponRepository.java
@@ -4,8 +4,8 @@ import cart.dao.CouponDao;
 import cart.dao.MemberCouponDao;
 import cart.dao.MemberDao;
 import cart.domain.Member;
-import cart.domain.MemberCoupon;
 import cart.domain.coupon.Coupon;
+import cart.domain.coupon.MemberCoupon;
 import cart.domain.coupon.policy.DiscountPolicy;
 import cart.domain.coupon.policy.DiscountPolicyType;
 import cart.entity.CouponEntity;
