+++ src/main/java/cart/repository/OrderRepository.java
@@ -6,10 +6,10 @@ import cart.dao.MemberDao;
 import cart.dao.OrderItemDao;
 import cart.dao.OrdersDao;
 import cart.domain.Member;
-import cart.domain.MemberCoupon;
 import cart.domain.Order;
 import cart.domain.OrderItem;
 import cart.domain.coupon.Coupon;
+import cart.domain.coupon.MemberCoupon;
 import cart.domain.coupon.policy.DiscountPolicy;
 import cart.domain.coupon.policy.DiscountPolicyType;
 import cart.entity.CouponEntity;
